Add new version of AWPMD packet
The new version of AWPMD packet was presented. This version include new pair
styles for WPMD with Antisimmetrization and Hartree approaches. Also
the WPMD with DFT extension was introduces as a set of pair styles.

Also, the next fixes were added:
 * fix for Monte-Carlo energy minimization
 * fix for harmonic trap for wavepacket atom style
 * fix for nve/nvh ansamble with WPMD support

Other changes:
 * change build instruction for AWPMD package. All build process performes
   with cmake
 * the documentation was updated with new information about WPMD/AWPMD
 * add additional examples for AWPMD package
 * add extera parameters for set/compute command

+++ lib/awpmd/awpmd/awpmd-dft/nvgpu/awpmd-dft-nvgpu.cu
@@ -202,8 +202,8 @@ std::tuple<double, double> XCEnergy_nvgpu::adaptive_mesh_integration(std::vector
     compute_forces(derivatives, wavepackets, energy_eval);
   }
 
-  return std::make_tuple<double, double>(thrust::get<0>(xck_energy),
-                                         thrust::get<1>(xck_energy));
+  return std::make_tuple<double, double>((double)thrust::get<0>(xck_energy),
+                                         (double)thrust::get<1>(xck_energy));
 }
 
 std::tuple<double, double> XCEnergy_nvgpu::regular_mesh_integration(
@@ -224,8 +224,8 @@ std::tuple<double, double> XCEnergy_nvgpu::regular_mesh_integration(
     compute_forces(derivatives, wavepackets, energy_eval);
   }
 
-  return std::make_tuple<double, double>(thrust::get<0>(xck_energy),
-                                         thrust::get<1>(xck_energy));
+  return std::make_tuple<double, double>((double)thrust::get<0>(xck_energy),
+                                         (double)thrust::get<1>(xck_energy));
 }
 
 void
